fix: move non-prerendered pages to components
